+++ stream-projects/chat-theme/routes/page-theme/chat-theme-dev-embed.json
@@ -21,5 +21,21 @@
     "iframeUrl": "https://local-chat-theme.rileymiller.dev/chat-theme",
     "orgId": "6f4edfe0-8446-11ec-b10d-3a946723cad5",
     "slug": "https://local-chat-theme.rileymiller.dev/chat-theme"
+  },
+  {
+    "id": "https://local-chat-theme.rileymiller.dev/page-theme",
+    "defaultLayoutConfigSteps": [
+      {
+        "action": "querySelector",
+        "value": "ytd-app"
+      },
+      {
+        "action": "appendSubject"
+      }
+    ],
+    "domAction": "append",
+    "iframeUrl": "https://local-chat-theme.rileymiller.dev/page-theme",
+    "orgId": "6f4edfe0-8446-11ec-b10d-3a946723cad5",
+    "slug": "https://local-chat-theme.rileymiller.dev/page-theme"
   }
 ]
