+++ stream-projects/chat-theme/routes/page-theme/page.tsx
@@ -1,0 +1,9 @@
+import React from "https://npm.tfl.dev/react";
+import { toDist } from "https://tfl.dev/@truffle/distribute@^2.0.0/format/wc/react/index.ts"; // DO NOT BUMP;
+import PageTheme from "../../components/page-theme/page-theme.tsx";
+
+function PageThemePage() {
+  return <PageTheme />;
+}
+
+export default toDist(PageThemePage, import.meta.url);
