+++ stream-projects/chat-theme/shared/mod.ts
@@ -1,2 +1,3 @@
 export * from "./extension/mod.ts";
 export * from './alerts/mod.ts';
+export * from './themes/mod.ts';
